Add afl-gcc-fast for plugin-based instrumentation
This adds support for GCC-plugin-powered instrumentation, through a new
`afl-gcc-fast` program, meaning that gcc-powered fuzzing is now also available
on non-X86 platforms.[1] This pass is initialized after GCC has SSA-ified the
input (which happens rather early in the compilation process), by inserting
GIMPLE into each basic block for the bitmap update.

The current instrumentation method is undoubtly slower than afl-clang-fast,
although there's no fundamental reason for it to be so. Currently, rather than
inlining the basic block trace directly into the header of each block, we
instead emit GIMPLE for a 'call' to an __afl_trace function (in afl-gcc-rt.o.c)
with the current location as input, and it does the bitmap update. Obviously,
this implementation does get perfect branch coverage since it is precise,
clearly.

The reason for this is mostly a boring one: GCC has very, very little
documentation available, and it's proven rather difficult to determine the
proper structure of MEM_REF nodes for doing pointer dereferences. You'd think
this would be easy. I'm probably missing something simple; most of the
(currently unused) code is still available inline.

In any case, that will be fixed soon enough hopefully, meaning afl-gcc-fast
will essentially be completely at-parity with afl-clang-fast (and the inline
bitmap update should come out to about 4-5 instructions on x86.)

The internal structure is almost identical to afl-clang-fast, modulo a few
things. All things considered, these could all be bundled up into the same tool
with a few extra 'if' clauses, and afl-gcc-rt and afl-llvm-rt could be
collapsed together as well. A future exercise.

Overall, the instrumentation works exactly the same as `afl-clang-fast` from a
user perspective.

To build, simply do:

    $ make PREFIX=<FOO>
    $ cd gcc_plugin && make PREFIX=<FOO>
    $ cd ..
    $ make install PREFIX=<FOO>

You'll now have 'afl-gcc-fast' and 'afl-g++-fast' available to fuzz with; on
Debian machines you'll need a gcc-<version>-plugin-dev package to compile it.

Signed-off-by: Austin Seipp <aseipp@pobox.com>